Add support for forwarding-and-encrypting zephyrs
If you have a server process (or users) that should be able to send you
zephyrs, but you want to send them to an encrypted class and don't want
to deal with key management, this may be useful. It lets your server
just send a regular personal zephyr, and the forwarder will encrypt it
and pass it along.