Add support for auto-connecting to Jabber accounts
This adds a new config key, ClassPair->jabber->autoconnect. When set to true,
the module will automatically use the supplied password to connect to the
Jabber account for this ClassPair.

This also changes the support method for reloading config to
:forwarder:load_config (instead of :reload-module Forwarder). The
:reload-module meant that we weren't using the startup hook, which was
problematic for using Jabber commands in the module's setup functionality.

This also incidentally adds the warn() function.